feat(ui): rename Menu component to MenuCategory
