+++ digimenu-frontend/src/layouts/components/CategoryMenu/index.js
@@ -1,0 +1,1 @@
+export { default } from './CategoryMenu';
